+++ R/plot_qsm.R
@@ -66,6 +66,10 @@ plot_qsm <- function(
     }
   }
 
+  # User selected columns
+  radius <- select_column(rlang::enquo(radius))
+  color <- select_column(rlang::enquo(color))
+
   # rTwig ----------------------------------------------------------------------
   if (all(c("id", "parent", "start_x", "branch_order") %in% colnames(cylinder))) {
     plot_data(