+++ NEWS.md
@@ -4,14 +4,17 @@
 
 -   aRchi QSMs are now supported in the package.
 
-<!-- -->
+-   Users can now pass column variables either unquoted or quoted to user facing functions like `plot_qsm()` or `qsm_summary()`. Piping the QSM cylinders into these functions allows the user to easily select unquoted columns.
 
 -   `plot_qsm()`: Point clouds can now be plotted independently of a QSM.
+
     -   *pan*: The right mouse button now pans the plot instead of zooming by default. The functionality mimics `pan3d()` from RGL, but rewritten with Rcpp for better responsiveness.
     -   *lit*: Enable or disable plot lighting
     -   *axes_color*: The axes color can now be set by the user.
     -   *skeleton_lwd*: Set the line width of the skeleton plots.
+
 -   `correct_radii()`: The broken branch filter can now be disabled, with *broken_branch = FALSE*
+
 -   `run_rtwig()`: The broken branch filter can now be disabled, with *broken_branch = FALSE*
 
 ## Improvements